requirements.txt
alabaster==0.7.16
asgiref==3.11.1
babel==2.18.0
certifi==2026.2.25
charset-normalizer==3.4.5
defusedxml==0.7.1
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
docutils==0.21.2
idna==3.11
imagesize==1.5.0
importlib_metadata==8.7.1
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.1.1
oauthlib==3.3.1
packaging==26.0
pillow==11.3.0
Pygments==2.20.0
requests==2.32.5
requests-oauthlib==2.0.0
snowballstemmer==3.0.1
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
